pyproject.toml
[tool.poetry]
name = "synthetic-loan-example"
version = "0.1.0"
description = ""
authors = ["theovald <gjoegjoe@outlook.com>"]
readme = "README.md"
packages = [{include = "synthetic_data"}]

[tool.poetry.dependencies]
python =">=3.10,<3.11"
fairlearn = "^0.8.0"
matplotlib = "^3.6.2"
ipy_table = "^1.15.1"
dataframe-image = "^0.1.3"
seaborn = "^0.12.2"
pandas-profiling = "^3.6.2"
jupyter-black = "^0.3.3"
autoflake = "^2.0.0"
jupyter = "^1.0.0"
nbstripout = "^0.6.1"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
